F1

+++ Controllers/CareLogController.cs
@@ -8,16 +8,38 @@ namespace HomePlant.Controllers;
 public class CareLogController : Controller
 {
     private readonly CareLogService _service;
+    private readonly UserPlantService _userPlantService;
 
     public CareLogController(
-        CareLogService service)
+        CareLogService service,
+        UserPlantService userPlantService)
     {
         _service = service;
+        _userPlantService = userPlantService;
+    }
+
+    private async Task<IActionResult?> RequireOwnedPlant(
+        string plantId)
+    {
+        var uid = HttpContext.Session.GetString("Uid");
+
+        if (uid == null)
+            return RedirectToAction("Login", "Account");
+
+        var plant = await _userPlantService.GetById(uid, plantId);
+
+        if (plant == null)
+            return NotFound();
+
+        return null;
     }
 
     public async Task<IActionResult> Index(
         string plantId)
     {
+        if (await RequireOwnedPlant(plantId) is IActionResult redirect)
+            return redirect;
+
         ViewBag.PlantId = plantId;
 
         var logs =
@@ -26,9 +48,12 @@ public class CareLogController : Controller
         return View(logs);
     }
 
-    public IActionResult Create(
+    public async Task<IActionResult> Create(
         string plantId)
     {
+        if (await RequireOwnedPlant(plantId) is IActionResult redirect)
+            return redirect;
+
         ViewBag.PlantId = plantId;
 
         return View();
@@ -39,6 +64,9 @@ public class CareLogController : Controller
         string plantId,
         CareLogModel model)
     {
+        if (await RequireOwnedPlant(plantId) is IActionResult redirect)
+            return redirect;
+
         model.CreatedAt =
             Timestamp.GetCurrentTimestamp();
 
@@ -55,6 +83,9 @@ public class CareLogController : Controller
         string plantId,
         string id)
     {
+        if (await RequireOwnedPlant(plantId) is IActionResult redirect)
+            return redirect;
+
         await _service.DeleteLog(
             plantId,
             id);
